First Setup System
DiskIndex now checks if the rootFolderPath.json is correctly configured, if its empty it will activate a dialog pop-up to enter the directory.

The system will check if the directory inserted in the dialog exists otherwise it will not apply the changes; if the directory is correct and the User clicks "Apply" the system will automatically write it in the rootFolderPath.json without the need of manually modifying the file.

Its still possible to escape the pop-up with a "Cancel" button.